requirements.txt
requests~=2.24.0
Flask~=1.1.2
flask-sqlalchemy~=2.4.4